file management and norminette friendly
